chore: restore from QR code image & video
